+++ ios/Framework/Core/Auth/Common/AuthType.swift
@@ -17,6 +17,8 @@ import UIKit
 
 	case basic = 1
 	case cookie = 2
+	case oauth2Redirect = 3
+	case oauth2UsernameAndPassword = 4
 
 }
 
@@ -26,6 +28,10 @@ public func StringFromAuthType(_ authType: AuthType) -> String {
 		return "basic"
 	case .cookie:
 		return "cookie"
+	case .oauth2Redirect:
+		return "oauth2redirect"
+	case .oauth2UsernameAndPassword:
+		return "oauth2usernameandpassword"
 	}
 }
 
@@ -35,6 +41,10 @@ public func AuthTypeFromString(_ str: String) -> AuthType? {
 		return .basic
 	case "cookie":
 		return .cookie
+	case "oauth2usernameandpassword":
+		return .oauth2UsernameAndPassword
+	case "oauth2redirect":
+		return .oauth2Redirect
 	default:
 		return nil
 	}