+++ ios/Framework/Themes/Default/Auth/LoginScreenlet/LoginView_default.swift
@@ -27,6 +27,12 @@ open class LoginView_default: BaseScreenletView, LoginViewModel {
 	@IBOutlet open weak var authorizeButton: UIButton?
 
 	override open var progressMessages: [String: ProgressMessages] {
+		let auth = AuthTypeFromString(authType ?? "basic")
+
+		if auth == .oauth2Redirect {
+			return [:]
+		}
+
 		return [
 			BaseScreenlet.DefaultAction: [
 				.working: LocalizedString("default", key: "login-loading-message", obj: self),
@@ -97,12 +103,10 @@ open class LoginView_default: BaseScreenletView, LoginViewModel {
 
 	override open func onStartInteraction() {
 		loginButton?.isEnabled = false
-		authorizeButton?.isEnabled = false
 	}
 
 	override open func onFinishInteraction(_ result: Any?, error: NSError?) {
 		loginButton?.isEnabled = true
-		authorizeButton?.isEnabled = true
 	}
 
 	override open func createProgressPresenter() -> ProgressPresenter {
@@ -118,6 +122,23 @@ open class LoginView_default: BaseScreenletView, LoginViewModel {
 	// MARK: Public methods
 
 	open func configureAuthType() {
-		_ = AuthTypeFromString(authType ?? "") ?? .basic
+		let auth = AuthTypeFromString(authType ?? "") ?? .basic
+
+		if auth == .oauth2Redirect {
+			showAuthorizeButton()
+		}
+		else {
+			showUserAndPasswordForm()
+		}
+	}
+
+	open func showAuthorizeButton() {
+		authorizeButton?.isHidden = false
+		loginButton?.superview?.isHidden = true
+	}
+
+	open func showUserAndPasswordForm() {
+		authorizeButton?.isHidden = true
+		loginButton?.superview?.isHidden = false
 	}
 }